- Fixed warnings when using int instead of NSInteger

+++ RichTextEditor/Source/Categories/NSAttributedString+RichTextEditor.m
@@ -41,7 +41,7 @@
 		range.location-1 :
 		range.location;
 	
-	for (int i=startingRange ; i>=0 ; i--)
+	for (NSInteger i=startingRange ; i>=0 ; i--)
 	{
 		char c = [self.string characterAtIndex:i];
 		if (c == '\n')
@@ -55,7 +55,7 @@
 	
 	NSInteger moveForwardIndex = (range.location > start) ? range.location : start;
 	
-	for (int i=moveForwardIndex; i<= self.string.length-1 ; i++)
+	for (NSInteger i=moveForwardIndex; i<= self.string.length-1 ; i++)
 	{
 		char c = [self.string characterAtIndex:i];
 		if (c == '\n')
